+++ src/main/java/com/example/ordermanager/owner/controller/OwnerController.java
@@ -7,6 +7,7 @@ import com.example.ordermanager.owner.service.OwnerManagementService;
 import com.example.ordermanager.payment.entity.Payment;
 import com.example.ordermanager.payment.service.PaymentService;
 import com.example.ordermanager.utils.PasswordVerificationService;
+import java.math.BigDecimal;
 import java.security.Principal;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
@@ -150,7 +151,8 @@ public class OwnerController {
 
   @PostMapping("/companies/{id}/set-fee")
   public String setCompanyFee(@PathVariable Long id,
-      @RequestParam(required = false) Double monthlyFee, RedirectAttributes redirectAttributes) {
+      @RequestParam(required = false) BigDecimal monthlyFee,
+      RedirectAttributes redirectAttributes) {
     try {
       companyService.setMonthlyFee(id, monthlyFee);
       redirectAttributes.addFlashAttribute("message", "Monthly fee updated successfully.");
